requirements.txt
tensorflow-macos==2.13.0
mediapipe==0.10.21
opencv-python==4.8.0.76
numpy==1.24.3
websockets==11.0.3
google-generativeai==0.8.5
python-dotenv==1.0.0
protobuf==4.25.3
typing-extensions==4.5.0 